kafkaProperties 변경

+++ User/src/main/java/com/areastory/user/kafka/ArticleListener.java
@@ -1,5 +1,6 @@
 package com.areastory.user.kafka;
 
+import com.areastory.user.config.properties.KafkaProperties;
 import com.areastory.user.dto.common.ArticleKafkaDto;
 import com.areastory.user.service.ArticleService;
 import lombok.RequiredArgsConstructor;
@@ -9,20 +10,19 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 public class ArticleListener {
+    private final KafkaProperties kafkaProperties;
     private final ArticleService articleService;
 
-    @KafkaListener(id = KafkaProperties.GROUP_NAME_ARTICLE, topics = KafkaProperties.TOPIC_ARTICLE, containerFactory = "articleContainerFactory")
+    @KafkaListener(id = "${kafka.group.article}", topics = "${kafka.topic.article}", containerFactory = "articleContainerFactory")
     public void articleListen(ArticleKafkaDto articleKafkaDto) {
-        switch (articleKafkaDto.getType()) {
-            case KafkaProperties.INSERT:
-                articleService.addArticle(articleKafkaDto);
-                break;
-            case KafkaProperties.UPDATE:
-                articleService.updateArticle(articleKafkaDto);
-                break;
-            case KafkaProperties.DELETE:
-                articleService.deleteArticle(articleKafkaDto);
-                break;
-        }
+        String type = articleKafkaDto.getType();
+        if (type.equals(kafkaProperties.getCommand().getInsert()))
+            articleService.addArticle(articleKafkaDto);
+
+        else if (type.equals(kafkaProperties.getCommand().getUpdate()))
+            articleService.updateArticle(articleKafkaDto);
+
+        else if (type.equals(kafkaProperties.getCommand().getDelete()))
+            articleService.deleteArticle(articleKafkaDto);
     }
 }
